+++ @MSIT/MSIT_Run_Trial.m
@@ -1,24 +1,22 @@
-function FixationStart = MSIT_Run_Trial(app, lastFixationStart)
+function MSIT_Run_Trial(app)
 
 % Runs a single trial of the neuropixels version of the MSIT.
 
 n = app.TaskSettings.CurrentTrial;
-FixationStart = NaN;
 
-%% Run ISI
-% If it's the first trial, show a fixation cross. Otherwise there will
-% already be a fixation cross up from the end of the previous trial
+%% Run first ISI
+% If it's the first trial, show a fixation cross and wait.
 if n == 1
-    lastFixationStart = MSIT_Show_Text(app, '+', 1);
-    Send_Log_TTL(app, lastFixationStart, 'Fixation_Shown', false, true);
+    FixationStart = MSIT_Show_Text(app, '+', 1);
+    Send_Log_TTL(app, FixationStart, 'Fixation_Shown', false, true);
+    % Wait out the ISI and check for escape key.
+    MSIT_Wait_And_Advance_Or_Cancel(app, FixationStart, app.TrialStruct.Trials(n).ISI, 'escape');
+    if app.ExitFlag
+        return;
+    end
 end
 
-% Wait out the ISI and check for escape key.
-MSIT_Wait_And_Advance_Or_Cancel(app, lastFixationStart, app.TrialStruct.Trials(n).ISI, 'escape');
-if app.ExitFlag
-    return;
-end
-
+%% Check if keys are down before presenting stimulus
 % Check if a key is down before moving on to present the stimulus.
 RestrictKeysForKbCheck(app.TaskSettings.ResponseCodes);
 keyIsDown = KbCheck;
@@ -29,9 +27,9 @@ while keyIsDown
     Send_Log_TTL(app, ShowTime, 'Key_Down_Message_Shown', false, true);
     KbWait([], 1);
     % Run the ISI again and send TTL
-    lastFixationStart = MSIT_Show_Text(app, '+', 1);
-    Send_Log_TTL(app, lastFixationStart, 'Fixation_Shown', false, true);
-    MSIT_Wait_And_Advance_Or_Cancel(app, lastFixationStart, app.TrialStruct.Trials(n).ISI, 'escape');
+    FixationStart = MSIT_Show_Text(app, '+', 1);
+    Send_Log_TTL(app, FixationStart, 'Fixation_Shown', false, true);
+    MSIT_Wait_And_Advance_Or_Cancel(app, FixationStart, app.TrialStruct.Trials(n).ISI, 'escape');
     if app.ExitFlag
         return;
     end
@@ -47,22 +45,27 @@ disp(['Displaying stimulus (' app.TrialStruct.Trials(n).Stimuli ') for trial ' n
 
 %% Get keyboard response
 MSIT_Wait_And_Get_Response(app, StimulusStart, app.TaskSettings.ResponseTime);
-% If we got a response, wait out the rest of the ResponseTime while
-% checking for exit key. If no response, report it
-if app.TrialStruct.Trials(n).ResponseKey == -1
-    disp('Response time exceeded, no response recorded.');
-else
-    MSIT_Wait_And_Advance_Or_Cancel(app, StimulusStart, app.TaskSettings.ResponseTime, 'escape');
-    if app.ExitFlag
-        return;
-    end
+% Wait out the rest of the ResponseTime while checking for exit key.
+MSIT_Wait_And_Advance_Or_Cancel(app, StimulusStart, app.TaskSettings.ResponseTime, 'escape');
+if app.ExitFlag
+    return;
 end
 
-%% Draw fixation cross for the next trial
+%% Draw fixation cross for the next trial, continue listening for keyboard response
 FixationStart = MSIT_Show_Text(app, '+', 1);
 Send_Log_TTL(app, FixationStart, 'Fixation_Shown', false, true);
-
-% If it's the last trial, let it run out its ISI
 if n == app.TaskSettings.nTrials
-    MSIT_Wait_And_Advance_Or_Cancel(app, FixationStart, app.TrialStruct.Trials(n).ISI, 'justwait');
+    isitime = app.TrialStruct.Trials(n).ISI;
+else
+    isitime = app.TrialStruct.Trials(n+1).ISI;
+end
+% If they haven't responded yet, check for late response
+exitkeychecktime = 0.05; % Reserve the last x seconds to check for exit key
+if app.TrialStruct.Trials(n).ResponseKey == -1
+    MSIT_Wait_And_Get_Response(app, StimulusStart, FixationStart-StimulusStart+isitime-exitkeychecktime);
+end
+% Wait out the rest of the isi checking for exit key
+MSIT_Wait_And_Advance_Or_Cancel(app, FixationStart, isitime, 'escape');
+if app.ExitFlag
+    return;
 end